Added tests to persistence layer.

+++ kypo2-commons-persistence/src/main/resources/db/migration/V1_1__create_db.sql
@@ -1,11 +1,11 @@
 CREATE TABLE idm_group_ref (
    id             BIGSERIAL NOT NULL,
-   group_id       BIGSERIAL NOT NULL UNIQUE,
+   group_id       int8 NOT NULL UNIQUE,
    PRIMARY KEY (id));
 
 CREATE TABLE user_ref (
    id             BIGSERIAL NOT NULL,
-   user_id       BIGSERIAL NOT NULL UNIQUE,
+   user_id       int8 NOT NULL UNIQUE,
    PRIMARY KEY (id));
 
 CREATE TABLE role (
